Fix mentor review: replace NavLink with Link, extract Tabs component
- Replace NavLink with Link in nav, use useLocation for is-active class
- Extract Tabs component to src/components/Tabs.tsx
- TabsPage now uses Tabs component with tabs data passed as props
