+++ src/App.tsx
@@ -1,51 +1,50 @@
 import 'bulma/css/bulma.css';
 import '@fortawesome/fontawesome-free/css/all.css';
 import './App.scss';
-import { NavLink, Navigate, Route, Routes } from 'react-router-dom';
+import { Link, Navigate, Route, Routes, useLocation } from 'react-router-dom';
 import { HomePage } from './pages/HomePage';
 import { TabsPage } from './pages/TabsPage';
 
-export const App = () => (
-  <>
-    <nav
-      className="navbar is-light is-fixed-top is-mobile has-shadow"
-      data-cy="Nav"
-    >
-      <div className="container">
-        <div className="navbar-brand">
-          <NavLink
-            to="/"
-            end
-            className={({ isActive }) =>
-              `navbar-item${isActive ? ' is-active' : ''}`
-            }
-          >
-            Home
-          </NavLink>
-          <NavLink
-            to="/tabs"
-            className={({ isActive }) =>
-              `navbar-item${isActive ? ' is-active' : ''}`
-            }
-          >
-            Tabs
-          </NavLink>
+export const App = () => {
+  const { pathname } = useLocation();
+
+  return (
+    <>
+      <nav
+        className="navbar is-light is-fixed-top is-mobile has-shadow"
+        data-cy="Nav"
+      >
+        <div className="container">
+          <div className="navbar-brand">
+            <Link
+              to="/"
+              className={`navbar-item${pathname === '/' ? ' is-active' : ''}`}
+            >
+              Home
+            </Link>
+            <Link
+              to="/tabs"
+              className={`navbar-item${pathname.startsWith('/tabs') ? ' is-active' : ''}`}
+            >
+              Tabs
+            </Link>
+          </div>
+        </div>
+      </nav>
+
+      <div className="section">
+        <div className="container">
+          <Routes>
+            <Route path="/" element={<HomePage />} />
+            <Route path="/home" element={<Navigate to="/" replace />} />
+            <Route path="tabs">
+              <Route index element={<TabsPage />} />
+              <Route path=":tabId" element={<TabsPage />} />
+            </Route>
+            <Route path="*" element={<h1 className="title">Page not found</h1>} />
+          </Routes>
         </div>
       </div>
-    </nav>
-
-    <div className="section">
-      <div className="container">
-        <Routes>
-          <Route path="/" element={<HomePage />} />
-          <Route path="/home" element={<Navigate to="/" replace />} />
-          <Route path="tabs">
-            <Route index element={<TabsPage />} />
-            <Route path=":tabId" element={<TabsPage />} />
-          </Route>
-          <Route path="*" element={<h1 className="title">Page not found</h1>} />
-        </Routes>
-      </div>
-    </div>
-  </>
-);
+    </>
+  );
+};
