+++ src/pages/TabsPage.tsx
@@ -1,5 +1,5 @@
-import { Link, useParams } from 'react-router-dom';
 import { Tab } from '../types/Tab';
+import { Tabs } from '../components/Tabs';
 
 const tabs: Tab[] = [
   { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
@@ -7,31 +7,9 @@ const tabs: Tab[] = [
   { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
 ];
 
-export const TabsPage = () => {
-  const { tabId } = useParams();
-  const activeTab = tabs.find(tab => tab.id === tabId);
-
-  return (
-    <>
-      <h1 className="title">Tabs page</h1>
-
-      <div className="tabs is-boxed">
-        <ul>
-          {tabs.map(tab => (
-            <li
-              key={tab.id}
-              data-cy="Tab"
-              className={tab.id === tabId ? 'is-active' : ''}
-            >
-              <Link to={`/tabs/${tab.id}`}>{tab.title}</Link>
-            </li>
-          ))}
-        </ul>
-      </div>
-
-      <div className="block" data-cy="TabContent">
-        {activeTab ? activeTab.content : 'Please select a tab'}
-      </div>
-    </>
-  );
-};
+export const TabsPage = () => (
+  <>
+    <h1 className="title">Tabs page</h1>
+    <Tabs tabs={tabs} />
+  </>
+);
